pyproject.toml
[project]
name = "aibox-cli"
# SINGLE SOURCE OF TRUTH: Version is read dynamically via importlib.metadata
version = "0.3.0"
description = "Container-Based Multi-AI Development Environment"
authors = [
    {name = "fogXploit"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["ai", "docker", "development", "claude", "gemini", "openai", "containers"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Software Distribution",
]

dependencies = [
    "typer>=0.12.0",           # CLI framework with type hints
    "rich>=13.7.0",            # Beautiful terminal output
    "docker>=7.0.0",           # Docker SDK for Python
    "pydantic>=2.6.0",         # Data validation using type hints
    "pyyaml>=6.0.1",           # YAML parser
    "questionary>=2.0.0",      # Interactive prompts
    "httpx>=0.27.0",           # Modern HTTP client
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",           # Testing framework
    "pytest-docker>=3.1.0",    # Docker fixtures for tests
    "pytest-cov>=4.1.0",       # Coverage reporting
    "pytest-mock>=3.12.0",     # Mocking support
    "mypy>=1.8.0",             # Static type checking
    "types-PyYAML>=6.0.0",     # Type stubs for PyYAML
    "ruff>=0.2.0",             # Fast linter and formatter
    "pre-commit>=3.6.0",       # Git hooks
]

[project.scripts]
aibox = "aibox.cli.main:app"

[project.urls]
Homepage = "https://github.com/fogXploit/aibox"
Documentation = "https://github.com/fogXploit/aibox/tree/main/docs"
Repository = "https://github.com/fogXploit/aibox"
Issues = "https://github.com/fogXploit/aibox/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aibox"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-docker>=3.1.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "ruff>=0.2.0",
    "pre-commit>=3.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes (includes F841: unused variables)
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear (includes B007: unused loop vars)
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
    "PIE", # flake8-pie (includes unused expressions)
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "docker.*",
    "questionary.*",
    "pydantic.*",
    "typer.*",
    "rich.*",
    "yaml",
]
ignore_missing_imports = true

# Allow subclassing Any for pydantic models and using untyped decorators
[[tool.mypy.overrides]]
module = [
    "aibox.config.models",
    "aibox.profiles.models",
    "aibox.cli.main",
    "aibox.cli.commands.*",
]
disallow_any_generics = false
disallow_subclassing_any = false
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: Slow tests",
    "unit: Unit tests",
]

[tool.coverage.run]
source = ["aibox"]
omit = [
    "tests/*",
    "**/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 75
